requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
# certifi @ file:///C:/ci/certifi_1655907258560/work/certifi
charset-normalizer==3.3.2
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
wincertstore==0.2
yarl==1.9.4
